split build graph json validation fixtures
